Expanded ParticleSystem and Improved Explosion Effect
Added a new method to generate a random direction vector in a plane from a rotation axis vector and a direction vector. With this we can create a nice Shockwave effect! 
Added Shockwave, Shrapnel and Smoke effects.
Also fixed the bug that only one type of particle was rendering at the same time.
Booom!