Implemented PackedJoinLobby
Changed return value off Lobby.addPLayer to String.
Errors are now split by "║" in data that is get send by lobbyPacket.
Filled Skeletons of PacketJoinLobby and PacketJoinLobbyStatus.
Added method isInLobby to abstract class Packet.
Added Variable curLObbbyId to Player. (this allows to find out if a Player is in a Lobby much faster)