Playernames get displayed in LobbyScreen
Added lobbyPlayerCatalog to game.class. (A list of players in a lobby and there status)
Changes in the update method of InLobby.class so that names ar e getting displayed.
If a CurrentLobbyInfo packet gets received by the client the player  names get saved in  lobbyPlayerCatalog.
A  LobbyOverviewPacket gets send now after  the Server registers a Disconnect.