Added a variable for the LobbyStatus "open", "running" or "finished"
It is no longer possible to join a Lobby witch is not open.
If a round of a lobby gets started its status is set to "running".
Lobbies that are not currently open are not shown in the LobbyOverview anymore.
added GameOver Methood to Lobby-Class which gets called from PlayerHandling.Player, if a certain amount of gold is reached.
Added PacketGameEnd Skeleton, which should be sent to the client from from the GameOvermethod in Lobby.