Its possible to create a Lobby in GUI
Added Stage LobbyCreation.
Added new Arrow Buttons for ChooseLobby-Menu.
Modified PacketCreateLobby for new data format (including LobbySize).
Added Textures for LobbyCreation-Menu.
Lobbycreation-Errors now get displayed in GUI.
It is now possible to select a mapsize when creating a Lobby.
(It is not possible yet to type in a Lobbyname, names are generated by Time in ms)
(It is not possible to select Mapsize via console, it uses always medium)