Implemented Ready
In the InLobby Menu the Ready-status gets displayed correctly.
The Match starts automatically as soon as all Lobbymembers are ready.
(This is also true if the only lobbymember that is not ready leaves the Lobby or disconnects)

Added Method allPlayersReady() to lobby class.
Added Method startRound() to lobby class.
The CurLobbyInfo-packet now also sends the Player-Ready status.
Added Variable ready to the Player  class.